+++ curso/aperitivos/templates/aperitivos/video.html
@@ -1,0 +1,72 @@
+{% load static %}
+<!DOCTYPE html>
+<html lang="pt-br">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <title>Rios de Luz</title>
+
+    <meta name="description" content="página do curso Rios de Luz">
+    <meta name="author" content="Sergio Casas">
+		<link rel="icon" type="image/png" href="{% static 'img/favicon.png' %}">
+    <link href="{% static 'css/bootstrap.min.css' %}" rel="stylesheet">
+	<link href="{% static 'css/style.css' %}" rel="stylesheet">
+    <script src="{% static 'js/jquery.min.js' %}"></script>
+    <script src="{% static 'js/bootstrap.min.js' %}"></script>
+    <script src="{% static 'js/scripts.js' %}"></script>
+
+  </head>
+  <body>
+	<nav class="navbar navbar-expand-lg navbar-light" style="background-color: #68c790;">
+		<a class="navbar-brand" href="{% url 'base:home' %}">Rios de Luz</a>
+		<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+		  <span class="navbar-toggler-icon"></span>
+		</button>
+	  
+		<div class="collapse navbar-collapse" id="navbarSupportedContent">
+		  <ul class="navbar-nav mr-auto">
+			<li class="nav-item">
+			  <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Disabled</a>
+			</li>
+		  </ul>
+		  
+		</div>
+	</nav>
+    <div class="container">
+		<div class="row">
+			<div class="col-md-12">
+                <h1 class="mt-4 mb-3">Videos Aperitivo: Motivação</h1>
+                <iframe src="https://player.vimeo.com/video/430007361" width="640" height="352" frameborder="0" allow="autoplay; fullscreen" allowfullscreen></iframe>
+			</div>
+		</div>
+	</div>
+
+<footer class='main-footer mt-5 pt-4 text-light'>
+	<div class="container">
+		<div class="row">
+			<div class="col">
+				<h3>Entre em contato</h3>
+				<address>
+					<a href="tel:+5512983010387">+55 12 98301-0387</a>
+					<br/>
+					<a href="mailto:contato@riosdeluz.brisaverde.life">contato@riosdeluz.brisaverde.life</a>
+				</address>
+
+			</div>
+		</div>
+	</div>
+	<div class="rights pb-4">
+		<div class="container">
+			<div class="row">
+				<div class="col">
+					{% now 'Y' %} Rios de Luz. Todos direitos reservados.
+				</div>
+			</div>
+		</div>
+	</div>
+</footer>
+
+  </body>
+</html>
