--- curso/aperitivos/templetes/aperitivos/video.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UFT-8">
-        <title>Title</title>
-    </head>
-    <body>
-    
-    </body>
-</html>
