--- curso/aperitivos/test/tests.py
@@ -1,12 +1,0 @@
-import pytest
-
-from django.urls import reverse
-
-
-@pytest.fixture
-def resp(client):
-    return client.get(reverse('aperitivos:video', args=('motivacao',)))
-
-
-def test_status_code(resp):
-    assert resp.status_code == 200
